Documents/Academlo/python-backend/proyectos/pythonProject/requirements.txt
Django==2.2.24
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
